Use custom tree picker instead of Umbraco's
Umbraco's built in tree picker assumes the data you're picking is an UmbracoEntity, which it may not be in this case.  It also has specific code for list view, search, etc that seemed like it may be difficult to re-use, so creating our own "tree picker" dialog for now.  We still use Umbraco's actual tree component within the dialog.  Note, Umbraco's tree also assumes you are rendering an "UmbracoTree" (with a section and all), so it may not work for our idea of a custom DotNet tree, unless we wrap it in an Umbraco tree.